requirements.txt
openai==0.28.1
python-dotenv==1.0.0
streamlit==1.27.1
metaphor-python